Move the test user to his own file
