+++ tests/Actions/CreateNewUserTest.php
@@ -2,15 +2,13 @@
 
 declare(strict_types=1);
 
-use ARKEcosystem\Fortify\Actions\CreateNewUser;
+use Tests\stubs\TestUser;
 use ARKEcosystem\Fortify\Models;
 
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Hash;
-use Spatie\MediaLibrary\MediaCollections\Models\Concerns\HasUuid;
+use Illuminate\Support\Facades\Config;
 use function Tests\expectValidationError;
+use ARKEcosystem\Fortify\Actions\CreateNewUser;
 
 beforeEach(function () {
     $this->validPassword = 'Pas3w05d&123456';
@@ -171,7 +169,7 @@ it('should require one special character', function () {
 
 it('handles the invitation parameter', function () {
     Config::set('fortify.models.user', \ARKEcosystem\Fortify\Models\User::class);
-    Config::set('fortify.models.invitation', CreateNewUserTest::class);
+    Config::set('fortify.models.invitation', TestUser::class);
 
     $user = (new CreateNewUser())->create([
         'name'                  => 'John Doe',
@@ -190,7 +188,7 @@ it('handles the invitation parameter', function () {
 
 it('marks the user email as verified if has an invitation', function () {
     Config::set('fortify.models.user', \ARKEcosystem\Fortify\Models\User::class);
-    Config::set('fortify.models.invitation', CreateNewUserTest::class);
+    Config::set('fortify.models.invitation', TestUser::class);
 
     $user = (new CreateNewUser())->create([
         'name'                  => 'John Doe',
@@ -220,35 +218,3 @@ it('doesnt mark the user email as verified if no ivitation ', function () {
     $this->assertNull($user->email_verified_at);
 });
 
-/**
- * @coversNothing
- */
-class CreateNewUserTest extends Model
-{
-    use HasUuid;
-
-    public ?string $uuid = null;
-
-    public ?int $user_id = null;
-
-    protected $guarded = [];
-
-    public static $model = null;
-
-    public static function findByUuid(string $uuid): ?Model
-    {
-        if (self::$model) {
-            return self::$model;
-        }
-
-        self::$model = new self(compact('uuid'));
-
-        return self::$model;
-    }
-
-    public function update(array $attributes = [], array $options = [])
-    {
-        $this->user_id = Arr::get($attributes, 'user_id', $this->user_id);
-        $this->uuid    = Arr::get($attributes, 'uuid', $this->uuid);
-    }
-}
